Added validation checks for board_handler

+++ internal/db/queries/boards.sql
@@ -1,7 +1,3 @@
--- name: GetAllBoards :many
-SELECT id, name, description, created_at, updated_at
-FROM boards;
-
 -- name: GetAllBoardsByUserId :many
 SELECT id, name, description, user_id, created_at, updated_at
 FROM boards
@@ -12,17 +8,20 @@ INSERT INTO boards (name, description, user_id)
 VALUES ($1, $2, $3)
 RETURNING id, name, description, user_id, created_at, updated_at;
 
--- name: DeleteBoardById :execrows
+-- name: DeleteBoardByIdAndUserId :execrows
 DELETE FROM boards
-WHERE id = $1;
-
--- name: GetBoardById :one
-SELECT id, name, description, created_at, updated_at 
-FROM boards
-WHERE id = $1;
-
--- name: UpdateBoardById :one
-UPDATE boards
-SET name = $2, description = $3, updated_at = CURRENT_TIMESTAMP
 WHERE id = $1
-RETURNING id, name, description, created_at, updated_at;
+AND user_id = $2;
+
+-- name: GetBoardByIdAndUserId :one
+SELECT id, name, description, user_id, created_at, updated_at 
+FROM boards
+WHERE id = $1
+AND user_id = $2;
+
+-- name: UpdateBoardByIdAndUserId :one
+UPDATE boards
+SET name = $3, description = $4, updated_at = CURRENT_TIMESTAMP
+WHERE id = $1
+AND user_id = $2
+RETURNING id, name, description, user_id, created_at, updated_at;
